+++ src/drawable.rs
@@ -1,7 +1,7 @@
-use crate::color::Color;
 use crate::shapes::CreateDrawable;
 use graphics_shapes::coord::Coord;
 use graphics_shapes::{IntersectsContains, Shape};
+use ici_files::prelude::*;
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};